+++ Core/printf/ft_printf.h
@@ -10,12 +10,12 @@
 int	ft_printf(const char *, ...);
 
 /* FORMATTING */
-void	ft_putchar(char c, int *cnt);
-void	ft_putstr(char *s, int *cnt);
-void	ft_putnbr(int n, int *cnt);
-void	ft_putunbr(unsigned int n, int *cnt);
-void	ft_puthex(unsigned int n, char *base, int *cnt);
-void	ft_putptr(void *ptr, int *cnt);
+void	ft_putchar(char c, size_t *cnt);
+void	ft_putstr(char *s, size_t *cnt);
+void	ft_putnbr(int n, size_t *cnt);
+void	ft_putunbr(unsigned int n, size_t *cnt);
+void	ft_puthex(unsigned int n, char *base, size_t *cnt);
+void	ft_putptr(void *ptr, size_t *cnt);
 
 /* HELPER */
 size_t	ft_strlen(const char *s);
